# dkquantweb/dkquantweb/OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "描述：用于管理集团旗下公司的人员信息,具体包括XXX,XXX模块...",
    "version": "版本号:3.8.9",
    "title": "标题：若依管理系统_接口文档",
    "contact": {
      "name": "RuoYi"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/dev-api/strategy/fun_fund/{id}": {
      "get": {
        "tags": [
          "资金方法相关接口"
        ],
        "summary": "获取资金方法详细信息",
        "operationId": "getInfoUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/fun_position/{id}": {
      "get": {
        "tags": [
          "开仓方法相关接口"
        ],
        "summary": "获取开仓方法详细信息",
        "operationId": "getInfoUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/fun_profitLoss/{id}": {
      "get": {
        "tags": [
          "止盈止损方法相关接口"
        ],
        "summary": "获取止盈止损方法详细信息",
        "operationId": "getInfoUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/fun_selection/{id}": {
      "get": {
        "tags": [
          "选股方法相关接口"
        ],
        "summary": "获取选股方法详细信息",
        "operationId": "getInfoUsingGET_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/strategy_fund/{id}": {
      "get": {
        "tags": [
          "资金策略相关接口"
        ],
        "summary": "获取资金策略详细信息",
        "operationId": "getInfoUsingGET_5",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/strategy_position/{id}": {
      "get": {
        "tags": [
          "开仓策略相关接口"
        ],
        "summary": "获取开仓策略详细信息",
        "operationId": "getInfoUsingGET_7",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/strategy_profitLoss/{id}": {
      "get": {
        "tags": [
          "止盈止损策略相关接口"
        ],
        "summary": "获取止盈止损策略详细信息",
        "operationId": "getInfoUsingGET_8",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/strategy_pv/{id}": {
      "get": {
        "tags": [
          "私有策略相关接口"
        ],
        "summary": "获取私有策略详细信息",
        "operationId": "getInfoUsingGET_9",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/strategy_selection/{id}": {
      "get": {
        "tags": [
          "选股策略相关接口"
        ],
        "summary": "获取选股策略详细信息",
        "operationId": "getInfoUsingGET_10",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/account/{id}": {
      "get": {
        "tags": [
          "策略账户信息相关接口"
        ],
        "summary": "获取策略账户信息详细信息",
        "operationId": "getInfoUsingGET_11",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/args/{id}": {
      "get": {
        "tags": [
          "策略参数信息相关接口"
        ],
        "summary": "获取策略参数信息详细信息",
        "operationId": "getInfoUsingGET_12",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/orders/{id}": {
      "get": {
        "tags": [
          "策略订单信息相关接口"
        ],
        "summary": "获取策略订单信息详细信息",
        "operationId": "getInfoUsingGET_13",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/position/{id}": {
      "get": {
        "tags": [
          "策略仓位相关接口"
        ],
        "summary": "获取策略仓位详细信息",
        "operationId": "getInfoUsingGET_14",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}